+++ attributes/domain_controllers.rb
@@ -1,0 +1,19 @@
+# N.California
+default['ad-nativex']['domain_controllers']['us-west-1'] = {
+    :primary => nil,
+    :backup => nil
+}
+# Oregon
+default['ad-nativex']['domain_controllers']['us-west-2'] = {
+    :primary => nil,
+    :backup => nil
+}
+# N.Virginia
+default['ad-nativex']['domain_controllers']['us-east-1'] = {
+    :primary => nil,
+    :backup => nil
+}
+default['ad-nativex']['domain_controllers']['on-premise'] = {
+    :primary => nil,
+    :backup => nil
+}